settings deleteBooks

+++ Backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Book = require("../models/Books");
+const fs = require("fs");
 
 exports.createBooks = (req, res, next) => {
     if (!req.body.book) {
@@ -39,9 +40,22 @@ exports.modifyBooks = (req, res, next) => {
 };
 
 exports.deleteBooks = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(book => res.status(200).json(book))
-        .catch(error => res.status(404).json({ error }));
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: "Non autorisé" });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: "Objet supprimé !" }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
 
 exports.readOneBook = (req, res, next) => {
